requirements.txt
aiopg==1.3.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.6.0
cffi==1.15.1
channels==3.0.5
channels-postgres==0.0.4
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
virtualenv==20.15.1
zope.interface==5.4.0
